refactor: Move VBE classes to their own files
